+++ dev/src/js/states/menu.js
@@ -20,16 +20,35 @@ var MenuState = (function (_super) {
         var button = Lightning.Geometry.Rect(50, 50);
         button.tint = 0xff22aa;
         this.add(button);
-        this.game.keyboard.key('a').subscribeOnce('pressed', function () {
-            console.log('I have been pressed mother fucker');
+        var sound = this.game.audio.load('meeseeks', ['audio.mp3']);
+        this.game.keyboard.key('space').subscribe('pressed', function () {
+            sound.play();
         });
         this.game.keyboard.key('a').subscribeOnce('pressed', function () {
-            console.log('here too?');
+            // console.log('I have been pressed mother fucker');
+        });
+        this.game.keyboard.key('a').subscribeOnce('pressed', function () {
+            // console.log('here too?');
         });
         button.interactive = true;
-        button.on('mousedown', function () {
+        button.on('pointerdown', function () {
+            console.log('yoo');
             _this.game.states.start('game');
         });
+        this.particleEmitter = new Lightning.ParticleEmitter(this, 0, this.game.center.y);
+        this.add(this.particleEmitter);
+        // generate a texture for the particle emitter
+        var texture = Lightning.Geometry.Rect(8, 8).generateCanvasTexture();
+        // add that texture to the particle emitter
+        this.particleEmitter.add(texture);
+        this.particleEmitter.enableDebug();
+        this.particleEmitter.setGravity(0, 0.01);
+        this.particleEmitter.setVelocityRange(-0.3, 0.3, -0.3, 0.3);
+        this.particleEmitter.setInterval(50);
+        this.particleEmitter.addToLocal = false;
+        this.particleEmitter.setRotationIncrement(-0.1, 0.1);
+        this.particleEmitter.setStrength(2);
+        this.particleEmitter.start();
     };
     MenuState.prototype.update = function () {
     };