+++ src/core/engine.ts
@@ -11,39 +11,57 @@ namespace Lightning {
          * @param {number} height
          * @param {string} canvasId
          */
-        constructor(width, height, wrapperId:string, options?:iEngineOptions) {
+        constructor(width:number, height:number, options?:iEngineOptions) {
             super();
-            this.initalise(width, height, wrapperId, this.prepareOptions(options));
+            this.initialize(width, height, this.prepareOptions(options));
         }
 
         private prepareOptions(options:iEngineOptions):iEngineOptions {
+            if(options.rendererOptions === undefined) options.rendererOptions = {};
+            if(options.plugins === undefined) options.plugins = {};
+
             let ops:iEngineOptions = {
                 skipHello: options.skipHello || true,
                 autoStart: options.autoStart || true,
                 renderer: options.renderer || 'auto',
-                resolution: options.resolution || 1, 
-                scaleManeger: options.scaleManeger || true, 
-                device: options.device || true,
-                storage: options.storage || true,
-                events: options.events || true,
-                keyboard: options.keyboard || true,
-                physicsLite: options.physicsLite || true,
-                box2d: options.box2d || false,
-                services: options.services || true,
-                sockets: options.sockets || false,
-                states: options.states || true,
-                tweens: options.tweens || true,
-                parallax: options.parallax || true,
-                particles: options.particles || true,
-                timer: options.timer || true,
-                webfonts: options.webfonts || true,
-                debug: options.debug || true,
-                maths: options.maths || true,
+                rendererOptions: {
+                    view: options.rendererOptions.view || null,
+                    transparent: options.rendererOptions.transparent || false,
+                    antialias: options.rendererOptions.antialias || false,
+                    preserveDrawingBuffer: options.rendererOptions.preserveDrawingBuffer || false,
+                    backgroundColor: options.rendererOptions.backgroundColor || 0xffffff,
+                    clearBeforeRender: options.rendererOptions.clearBeforeRender || true,
+                    forceCanvas: options.rendererOptions.forceCanvas || false,
+                    roundPixels: options.rendererOptions.roundPixels || false,
+                    forceFXAA: options.rendererOptions.forceFXAA || false,
+                    legacy: options.rendererOptions.legacy || false,
+                },
+                divID: options.divID || null,
+                plugins: {
+                    scaleManeger: options.plugins.scaleManeger || true, 
+                    device: options.plugins.device || true,
+                    storage: options.plugins.storage || true,
+                    events: options.plugins.events || true,
+                    keyboard: options.plugins.keyboard || true,
+                    physicsLite: options.plugins.physicsLite || true,
+                    box2d: options.plugins.box2d || false,
+                    services: options.plugins.services || true,
+                    sockets: options.plugins.sockets || false,
+                    states: options.plugins.states || true,
+                    tweens: options.plugins.tweens || true,
+                    parallax: options.plugins.parallax || true,
+                    particles: options.plugins.particles || true,
+                    timer: options.plugins.timer || true,
+                    webfonts: options.plugins.webfonts || true,
+                    debug: options.plugins.debug || true,
+                    maths: options.plugins.maths || true,
+                }
+                
             }
             return ops;
         }
 
-        private initalise(width, height, wrapperId, options:iEngineOptions) {
+        private initialize(width, height, options:iEngineOptions) {
             /**
              * Say Hello
              */
@@ -54,14 +72,14 @@ namespace Lightning {
             /**
              * Initalise the Scale Manager
              */
-            if(options.scaleManeger === null || options.scaleManeger === undefined || options.scaleManeger === true) {
+            if(options.plugins.scaleManeger === null || options.plugins.scaleManeger === undefined || options.plugins.scaleManeger === true) {
                 this._scaleManager = new Scale(this, width, height, 0);
             }
             
             /**
              * Initalise the Device
              */
-            if(options.device === null || options.device === undefined || options.device === true) {
+            if(options.plugins.device === null || options.plugins.device === undefined || options.plugins.device === true) {
                 this._device = new Device(this);
             }
             
@@ -70,22 +88,24 @@ namespace Lightning {
              */
             // setup the canvas
             let wrapper = document.createElement('div');
-            wrapper.id = wrapperId || '';
+            wrapper.id = options.divID || '';
 
             document.body.appendChild(wrapper);
 
             /**
              * Initalise the renderer
              */
+            options.rendererOptions.resolution = options.rendererOptions.resolution || this._scaleManager.devicePixelRatio;
+            
             if(options.renderer === 'auto') {
-                this._renderer = PIXI.autoDetectRenderer(width, height, {resolution: this._scaleManager.devicePixelRatio});
+                this._renderer = PIXI.autoDetectRenderer(width, height, options.rendererOptions);
             } else if(options.renderer === 'canvas') {
-                this._renderer = new PIXI.CanvasRenderer(width, height, {resolution: this._scaleManager.devicePixelRatio});
+                this._renderer = new PIXI.CanvasRenderer(width, height, options.rendererOptions);
             } else if (options.renderer === 'webgl') {
-                this._renderer = new PIXI.WebGLRenderer(width, height, {resolution: this._scaleManager.devicePixelRatio});
+                this._renderer = new PIXI.WebGLRenderer(width, height, options.rendererOptions);
             } else {
                 // fallback if renderer type isn't canvas / webgl / auto
-                this._renderer = PIXI.autoDetectRenderer(width, height, {resolution: this._scaleManager.devicePixelRatio});
+                this._renderer = PIXI.autoDetectRenderer(width, height, options.rendererOptions);
             }
 
             wrapper.appendChild(this._renderer.view);
@@ -114,49 +134,49 @@ namespace Lightning {
             /**
              * Initalise the Storage Manager
              */
-            if(options.storage === null || options.storage === undefined || options.storage === true) {
+            if(options.plugins.storage === null || options.plugins.storage === undefined || options.plugins.storage === true) {
                 this._storageManager = new StorageManager();
             }
 
             /**
              * Initalise the Global Event Emitter
              */
-            if(options.events === null || options.events === undefined || options.events === true) {
+            if(options.plugins.events === null || options.plugins.events === undefined || options.plugins.events === true) {
                 this._eventEmitter = new EventEmitter();
             }
             
             /**
              * Initalise the Keyboard Manager
              */
-            if(options.keyboard === null || options.keyboard === undefined || options.keyboard === true) {
+            if(options.plugins.keyboard === null || options.plugins.keyboard === undefined || options.plugins.keyboard === true) {
                 this._keyboardManager = new KeyboardManager(this);
             }
 
             /**
              * Initalise the Lite Physics
              */
-            if(options.physicsLite === null || options.physicsLite === undefined || options.physicsLite === true) {
+            if(options.plugins.physicsLite === null || options.plugins.physicsLite === undefined || options.plugins.physicsLite === true) {
                 this._physicsLite = new LitePhysicsManager(this);
             }
 
             /**
              * Initalise the Service Manager
              */
-            if(options.services === null || options.services === undefined || options.services === true) {
+            if(options.plugins.services === null || options.plugins.services === undefined || options.plugins.services === true) {
                 this._serviceManager = new ServiceManager(this); 
             }
 
             /**
              * Initalise the State Manager
              */
-            if(options.states === null || options.states === undefined || options.states === true) {
+            if(options.plugins.states === null || options.plugins.states === undefined || options.plugins.states === true) {
                 this._stateManager = new StateManager(this);
             }
 
             /**
              * Initalise the State Manager
              */
-            if(options.tweens === null || options.tweens === undefined || options.tweens === true) {
+            if(options.plugins.tweens === null || options.plugins.tweens === undefined || options.plugins.tweens === true) {
                 this._tweenManager = new TweenManeger(this);
             }
 
@@ -166,6 +186,9 @@ namespace Lightning {
             if(options.autoStart === null || options.autoStart === undefined || options.autoStart === true) {
                 this.start();
             }
+
+            // need to wrap in options
+            this._audioManager = new AudioManager(this);
         }
 
         /**
@@ -205,7 +228,7 @@ namespace Lightning {
  * TODOS
  * Implement some sort of socket connectivity manager
  * Write some nice transitions for the state manager
- * Implement an animatins class for extending pixi animations
+ * Implement an animations class for extending pixi animations
  * Move enableDrag function to the display object
  * Particle emitter clear pool
  * Particle emitter add to world instead of child of the emitter
@@ -213,8 +236,8 @@ namespace Lightning {
  * Think about how to implement a light sprite for particles so they dont take up so much performance. It sucks on safari!
  * Particle emitter make a pre-create class that lets you store pooled sprited before the state is started
  * Think about making a debug module that's a container in it's own right. It should accept x number of text values
- *   and sort through them accordinly, ensuring nothing is ever overlapped
- * Explore the posibility of using light ray casting?
+ *   and sort through them accordingly, ensuring nothing is ever overlapped
+ * Explore the possibility of using light ray casting?
  * Particle emitter presets??
  * Build a webfont loader
  * Think about how best to implement some kind of camera system
@@ -226,5 +249,7 @@ namespace Lightning {
  * 2. Comprehensive Scale Manager
  * 3. Re-configure how states work
  *      1. Not happy with having to call funtions when manually overriding them
+ * 4. Allow particles to be added to stage, not to particle emitter... or to take into consideration their respective positions
+ * 
  * 
  */